Upgrade lint and move test tsconfig to proper spot
