+++ packages/fastify-renderer/package.json
@@ -81,7 +81,7 @@
     "@swc/jest": "^0.2.29",
     "@types/connect": "^3.4.35",
     "@types/jest": "^26.0.24",
-    "@types/node": "^14.17.20",
+    "@types/node": "^18.11.9",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.11",
     "@types/sanitize-filename": "^1.6.3",
@@ -91,13 +91,12 @@
     "fastify": "^3.29.0",
     "gitpkg": "^1.0.0-beta.2",
     "jest": "^28.1.3",
-    "jest-playwright-preset": "^2.0.0",
     "npm-run-all": "^4.1.5",
     "pino-pretty": "^4.8.0",
     "react": "0.0.0-experimental-4ead6b530",
     "react-dom": "0.0.0-experimental-4ead6b530",
     "rimraf": "^3.0.2",
-    "typescript": "^4.7.4"
+    "typescript": "^5.2.2"
   },
   "files": [
     "client",
