+++ packages/test-apps/simple-react/NavigationHistoryTest.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link, useLocation } from 'fastify-renderer/client/react'
 
 const NavigationHistoryTest = () => {
-  const [path, navigate] = useLocation()
+  const [_path, navigate] = useLocation()
 
   return (
     <>
