+++ packages/test-apps/simple-react/package.json
@@ -20,6 +20,12 @@
     "@playwright/test": "^1.39.0",
     "@swc/core": "^1.3.95",
     "@swc/jest": "^0.2.29",
-    "html-validator": "^5.1.18"
+    "@types/jest": "^26.0.24",
+    "@types/node": "^18.11.9",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.11",
+    "html-validator": "^5.1.18",
+    "jest-playwright-preset": "^2.0.0",
+    "typescript": "^5.2.2"
   }
 }
