+++ packages/test-apps/simple-react/test/switching-contexts.spec.ts
@@ -1,5 +1,5 @@
 import { Page } from 'playwright-chromium'
-import { newTestPage, reactReady, rootURL } from '../../helpers'
+import { newTestPage, reactReady, rootURL } from '../helpers'
 
 describe('navigation details', () => {
   let page: Page
