--- packages/test-apps/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["."],
-  "exclude": ["**/node_modules", "**/dist/**"],
-  "compilerOptions": {
-    "target": "ES2019",
-    "outDir": "../../node",
-    "module": "commonjs",
-    "lib": ["ES2020", "dom"],
-    "sourceMap": true,
-    "types": ["jest", "node", "jest-playwright-preset", "expect-playwright"]
-  }
-}
